Removed angular. Added React, Babel, Webpack. Architecture changes
