+++ .gitignore
@@ -1,5 +1,4 @@
-server/node_modules
-server/npm-debug.log
-npm-debug.log
+**/node_modules
+**/npm-debug.log
 
-client/dist/components
+public/bundle.js
