+++ app/app.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+ReactDOM.render(
+  <h1>Boilerplate</h1>,
+  document.getElementById('app')
+);
