--- client/.bowerrc
@@ -1,3 +1,0 @@
-{
-	"directory": "dist/components"
-}
