--- gulpfile.js
@@ -1,4 +1,0 @@
-var gulp = require('gulp');
-
-gulp.task('mocha', function() {
-});
