+++ public/index.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <title>Boilerplate</title>
+  </head>
+  <body>
+    <div id="app"></div>
+
+    <script src="bundle.js"></script>
+  </body>
+</html>
