--- server/database/models/actionModel.js
@@ -1,1 +1,0 @@
-var mongoose = require('mongoose');var Schema   = mongoose.Schema;var actionSchema = new Schema({	'name' : String,	'params' : Array,	'user_id' : {	 	type: Schema.Types.ObjectId,	 	ref: '{ref}'	}});module.exports = mongoose.model('Action', actionSchema);
